d20-stat-tracker
In my tabltetop games I like playing bards and spellcasters who buff themselves/others in martial combat but it was taking way too much time to track all of the stat changes and figure out what stacked. This tracks buffs and penalties to to-hit and damage and orginally looked like a spreadsheet, wanted to try making it pretty and user-friendly enough to share with other players via git-pages. It's my first 'real' project that I hoped to be helpful for anyone but myself so any and all advice would be well appreciated.